javaspring-buoi38

+++ src/main/java/vn/scrip/buoi38_bvn/repository/BookRepository.java
@@ -1,9 +1,7 @@
 package vn.scrip.buoi38_bvn.repository;
-
-
 import vn.scrip.buoi38_bvn.entites.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-
-public interface BookRepository extends JpaRepository<Book, Long> {
+public interface BookRepository extends JpaRepository<Book, Integer> {
 }
+
